[form-factors] Move 1/2^+->1/2^+ descriptions to new file
